+++ src/data/HumidityData.java
@@ -8,7 +8,7 @@ import Sensors.Humidity;
 public class HumidityData {
 
 	private long id;
-	private long value;
+	private double value;
 
 	public HumidityData() {
 	}
@@ -21,11 +21,11 @@ public class HumidityData {
 		this.id = id;
 	}
 
-	public long getValue() {
+	public double getValue() {
 		return value;
 	}
 
-	public void setValue(long value) {
+	public void setValue(double value) {
 		this.value = value;
 	}
